LayoutTests/fast/selectors/matches-complex-expected.txt

Check the complex selectors with :matches pseudo class, nested and multiple :matches.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "testcase:matches(:matches(bb, gg) > testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "testcase:matches(:matches(dd, cc) + *)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd, cc) + *)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd, cc) + *)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd, cc) + *)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(:matches(ee > ee > ee > ff > ee > ee > ff > gg) > *)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(ee > ee > ee > ff > ee > ee > ff > gg) > *)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(ee > ee > ee > ff > ee > ee > ff > gg) > *)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing "testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "testcase:matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb > testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb > testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb > testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(dd + dd) + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd + dd) + cc + dd ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd + dd) + cc + dd ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(dd + dd + cc + *) ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd + dd + cc + *) ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd + dd + cc + *) ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(:matches(:matches(:matches(dd) + dd) + *) + dd) ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(:matches(:matches(dd) + dd) + *) + dd) ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(:matches(:matches(dd) + dd) + *) + dd) ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(dd + * + :matches(cc + dd)) ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd + * + :matches(cc + dd)) ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd + * + :matches(cc + dd)) ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd + * + :matches(cc + dd)) ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(:matches(* + dd + cc + :matches(cc + dd)) ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(* + dd + cc + :matches(cc + dd)) ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(* + dd + cc + :matches(cc + dd)) ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(* + dd + cc + :matches(cc + dd)) ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(* + :matches(* + cc) + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* + :matches(* + cc) + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(* + :matches(* + cc) + dd ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* + :matches(* + cc) + dd ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(cc + * + * + :matches(cc + cc + cc) + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + * + * + :matches(cc + cc + cc) + dd ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + * + * + :matches(cc + cc + cc) + dd ~ testcase)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(cc + cc + cc + :matches(cc + :matches(unknown, *)) ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + :matches(cc + :matches(unknown, *)) ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + :matches(cc + :matches(unknown, *)) ~ testcase)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(:matches(:matches(* + * + * + *) + *, :matches(*) + cc + * + * + *) ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(* + * + * + *) + *, :matches(*) + cc + * + * + *) ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(* + * + * + *) + *, :matches(*) + cc + * + * + *) ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(* + * + * + *) + *, :matches(*) + cc + * + * + *) ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(:matches(:matches(aa > aa > aa > *) testcase, * + * + * + * + * ~ testcase))"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(aa > aa > aa > *) testcase, * + * + * + * + * ~ testcase))').length is 3
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(aa > aa > aa > *) testcase, * + * + * + * + * ~ testcase))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(aa > aa > aa > *) testcase, * + * + * + * + * ~ testcase))')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:matches(aa > aa > aa > *) testcase, * + * + * + * + * ~ testcase))')[2].id is "testcase3"
PASS styledElements is ["testcase1","testcase2","testcase3"]

Testing "testcase:matches(:visited)"
PASS document.querySelectorAll('#test-root testcase:matches(:visited)').length is 0
PASS styledElements is []

Testing "testcase:matches(cc + *, :visited)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + *, :visited)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + *, :visited)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(bb, gg) > testcase):matches(ff > gg > testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase):matches(ff > gg > testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase):matches(ff > gg > testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(bb, gg) > testcase):matches(ff > gg > testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing "testcase:matches(:matches(dd, cc) + *):matches(dd + * + *)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd, cc) + *):matches(dd + * + *)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(dd, cc) + *):matches(dd + * + *)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(ee > ee > ee > ff > ee > ee > ff > gg) > *):matches(testcase *)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(ee > ee > ee > ff > ee > ee > ff > gg) > *):matches(testcase *)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(ee > ee > ee > ff > ee > ee > ff > gg) > *):matches(testcase *)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing "testcase:matches(:matches(:not(dd)) + *)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:not(dd)) + *)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:not(dd)) + *)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(:not(cc)) + *)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:not(cc)) + *)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:not(cc)) + *)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(:matches(:not(cc), :not(dd)) + *)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:not(cc), :not(dd)) + *)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:not(cc), :not(dd)) + *)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(:not(cc), :not(dd)) + *)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase):matches(* + :matches(* + cc) + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase):matches(* + :matches(* + cc) + dd ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase):matches(* + :matches(* + cc) + dd ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase, * + :matches(* + cc) + dd ~ testcase):matches(* + :matches(* + cc) + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase, * + :matches(* + cc) + dd ~ testcase):matches(* + :matches(* + cc) + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase, * + :matches(* + cc) + dd ~ testcase):matches(* + :matches(* + cc) + dd ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(:matches(aa:matches(bb > aa > aa) > bb > aa) > aa > bb testcase, * + :matches(* + cc) + dd ~ testcase):matches(* + :matches(* + cc) + dd ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase):matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase):matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase):matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(bb > :matches(bb, bb > aa) > aa > bb testcase):matches(:matches(bb > aa > aa > bb > aa) > aa > bb testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:matches(aa + *, bb + *, cc + *, dd + *, ee + *, ff + *, gg + *, hh + *, ii + *, jj + *, kk + *, ll + *, mm + *, nn + *)"
PASS document.querySelectorAll('#test-root testcase:matches(aa + *, bb + *, cc + *, dd + *, ee + *, ff + *, gg + *, hh + *, ii + *, jj + *, kk + *, ll + *, mm + *, nn + *)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(aa + *, bb + *, cc + *, dd + *, ee + *, ff + *, gg + *, hh + *, ii + *, jj + *, kk + *, ll + *, mm + *, nn + *)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(aa + *, bb + *, cc + *, dd + *, ee + *, ff + *, gg + *, hh + *, ii + *, jj + *, kk + *, ll + *, mm + *, nn + *)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(aa > aa > aa > aa *, aa > aa > aa > dd *, aa > aa > aa > hh *, cc aa > aa > aa > hh *, cc ff > aa > aa > hh *, bb > aa > aa > bb *, aa > aa > aa > bb *, aa > cc > aa > aa *)"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > aa *, aa > aa > aa > dd *, aa > aa > aa > hh *, cc aa > aa > aa > hh *, cc ff > aa > aa > hh *, bb > aa > aa > bb *, aa > aa > aa > bb *, aa > cc > aa > aa *)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > aa *, aa > aa > aa > dd *, aa > aa > aa > hh *, cc aa > aa > aa > hh *, cc ff > aa > aa > hh *, bb > aa > aa > bb *, aa > aa > aa > bb *, aa > cc > aa > aa *)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > aa *, aa > aa > aa > dd *, aa > aa > aa > hh *, cc aa > aa > aa > hh *, cc ff > aa > aa > hh *, bb > aa > aa > bb *, aa > aa > aa > bb *, aa > cc > aa > aa *)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > aa *, aa > aa > aa > dd *, aa > aa > aa > hh *, cc aa > aa > aa > hh *, cc ff > aa > aa > hh *, bb > aa > aa > bb *, aa > aa > aa > bb *, aa > cc > aa > aa *)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > aa *, aa > aa > aa > dd *, aa > aa > aa > hh *, cc aa > aa > aa > hh *, cc ff > aa > aa > hh *, bb > aa > aa > bb *, aa > aa > aa > bb *, aa > cc > aa > aa *)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

PASS successfullyParsed is true

TEST COMPLETE
